Move ChecksumHandle and its methods into their own module
